OXDEV-7733 Cleanup test names and move to correct directories
